+++ src/main/java/de/samply/file/csv/writer/CsvWriterParameters.java
@@ -2,6 +2,7 @@ package de.samply.file.csv.writer;
 
 import de.samply.file.bundle.PathsBundle;
 import de.samply.file.csv.CsvRecordHeaderOrder;
+import java.nio.charset.Charset;
 
 public class CsvWriterParameters {
 
@@ -9,6 +10,9 @@ public class CsvWriterParameters {
   private PathsBundle pathsBundle;
   private String outputFilename;
   private Integer maxNumberOfRowsForFlush;
+  private String delimiter;
+  private String endOfLine;
+  private Charset charset;
 
   public Integer getMaxNumberOfRowsForFlush() {
     return maxNumberOfRowsForFlush;
@@ -42,4 +46,27 @@ public class CsvWriterParameters {
     this.outputFilename = outputFilename;
   }
 
+  public String getDelimiter() {
+    return delimiter;
+  }
+
+  public void setDelimiter(String delimiter) {
+    this.delimiter = delimiter;
+  }
+
+  public String getEndOfLine() {
+    return endOfLine;
+  }
+
+  public void setEndOfLine(String endOfLine) {
+    this.endOfLine = endOfLine;
+  }
+
+  public Charset getCharset() {
+    return charset;
+  }
+
+  public void setCharset(Charset charset) {
+    this.charset = charset;
+  }
 }