+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2022-07-06]
+## [1.0.0 - 2023-01-23]
 ### Added
 - Paths Bundle, Paths Bundle Manager
 - SLF4J with log4j
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - .dockerignore
 - Github build.yml
 - Fetch Bundle from Blaze Store
+- Config for csv delimiter, eol and charset
+- MUTATIONS_CSV_SCRIPT_INTERPRETER, NEW_FILES_DIRECTORY and PERSIST_DIRECTORY in Dockerfile
+- Apk upgrade in Dockerfile
 
 ### Changed
 - Create temporal root directory if it does not exist